% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions.R
\name{ars}
\alias{ars}
\title{Auxiliary function to set number of regimes and autoregressive orders}
\usage{
ars(nregim = 1, p = 1, q = 0, d = 0)
}
\arguments{
\item{nregim}{positive integer value indicating the number of regimes.}

\item{p}{a list with integer values indicating the autoregressive order within each regime}

\item{q}{a list with integer values indicating the maximum lags for the exogenous series within each regime}

\item{d}{a list with integer values indicating the maximum lags for the threshold series within each regime}
}
\description{
Auxiliary function to set number of regimes and autoregressive orders
}
