% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{coef.mtar}
\alias{coef.mtar}
\title{coef method for objects of class \code{mtar}}
\usage{
\method{coef}{mtar}(object, ..., FUN = mean)
}
\arguments{
\item{object}{an object of class \code{mtar} generated by a call to the function \code{mtar()}.}

\item{...}{optional arguments to FUN.}

\item{FUN}{the function to be applied to the chains resulting from the MCMC-type algorithm for
each model parameter. By default, \code{FUN} is set to \code{mean()}.}
}
\value{
a list composed of the values obtained from \code{FUN} when the latter is computed using
        chains resulting from the MCMC-type algorithm for each model parameter.
}
\description{
coef method for objects of class \code{mtar}
}
