% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesians.R
\name{DIC}
\alias{DIC}
\title{Deviance information criterion (DIC)}
\usage{
DIC(...)
}
\arguments{
\item{...}{one or several objects of the same class.}
}
\value{
A matrix with the values of the DIC for each object in the input.
}
\description{
This function computes the Deviance Information Criterion (DIC), which is an adjusted within-sample predictive accuracy measure.
}
\references{
Spiegelhalter D.J., Best N.G., Carlin B.P. and Van Der Linde A. (2002) Bayesian Measures of Model Complexity and Fit.
            Journal of the Royal Statistical Society Series B (Statistical Methodology), 64(4), 583–639.

Spiegelhalter D.J., Best N.G., Carlin B.P. and Van der Linde A. (2014). The deviance information criterion:
            12 years on. Journal of the Royal Statistical Society Series B (Statistical Methodology), 76(3), 485–493.
}
