% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperfunctions.R
\name{geweke.diagTAR}
\alias{geweke.diagTAR}
\title{Geweke's convergence diagnostic for objects of class \code{mtar}}
\usage{
geweke.diagTAR(x, frac1 = 0.1, frac2 = 0.5)
}
\arguments{
\item{x}{an object of class \code{mtar} generated by a call to the function \code{mtar()}}

\item{frac1}{fraction to use from beginning of chain}

\item{frac2}{fraction to use from end of chain}
}
\description{
This is a wrapper function that applies the routine \code{geweke.diag()} to the chains obtained from a call to the function \code{mtar()}.
}
\seealso{
geweke.diag
}
