% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperfunctions.R
\name{geweke.plotTAR}
\alias{geweke.plotTAR}
\title{Geweke-Brooks plot for objects of class \code{mtar}}
\usage{
geweke.plotTAR(
  x,
  frac1 = 0.1,
  frac2 = 0.5,
  nbins = 20,
  pvalue = 0.05,
  auto.layout = TRUE,
  ask,
  ...
)
}
\arguments{
\item{x}{an object of class \code{mtar} generated by a call to the function \code{mtar()}.}

\item{frac1}{fraction to use from beginning of chain}

\item{frac2}{fraction to use from end of chain}

\item{nbins}{Number of segments}

\item{pvalue}{p-value used to plot confidence limits for the null hypothesis}

\item{auto.layout}{If \code{TRUE} then, set up own layout for plots, otherwise use existing one}

\item{ask}{If \code{TRUE} then prompt user before displaying each page of plots. Default is \code{dev.interactive()}.}

\item{...}{Graphical parameters}
}
\description{
This is a wrapper function that applies the routine \code{geweke.plot()} to the chains obtained from a call to the function \code{mtar()}.
}
\seealso{
geweke.plot
}
