% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperfunctions.R
\name{geweke_plotTAR}
\alias{geweke_plotTAR}
\title{Geweke-Brooks plot for objects of class \code{mtar}}
\usage{
geweke_plotTAR(
  x,
  frac1 = 0.1,
  frac2 = 0.5,
  nbins = 20,
  pvalue = 0.05,
  auto.layout = TRUE,
  ask,
  ...
)
}
\arguments{
\item{x}{An object of class \code{mtar} returned by a call to \code{mtar()}.}

\item{frac1}{fraction to use from beginning of chain}

\item{frac2}{fraction to use from end of chain}

\item{nbins}{Number of segments}

\item{pvalue}{p-value used to plot confidence limits for the null hypothesis}

\item{auto.layout}{If \code{TRUE} then, set up own layout for plots, otherwise use existing one}

\item{ask}{If \code{TRUE} then prompt user before displaying each page of plots. Default is
\code{dev.interactive()}.}

\item{...}{Additional graphical parameters passed to the plotting routines.}
}
\description{
This function is a wrapper around \code{geweke.plot()} that applies the
Geweke-Brooks convergence diagnostic to the MCMC chains obtained from a
fitted \code{mtar} model.
}
\examples{
\donttest{
###### Example 1: Returns of the closing prices of three financial indexes
data(returns)
fit1 <- mtar(~ COLCAP + BOVESPA | SP500, data=returns, row.names=Date,
             subset={Date<="2015-12-07"}, dist="Student-t",
             ars=ars(nregim=3,p=c(1,1,2)), n.burnin=1000, n.sim=2000,
             n.thin=2)
geweke_plotTAR(fit1)

###### Example 2: Rainfall and two river flows in Colombia
data(riverflows)
fit2 <- mtar(~ Bedon + LaPlata | Rainfall, data=riverflows, row.names=Date,
             subset={Date<="2009-02-13"}, dist="Laplace",
             ars=ars(nregim=3,p=5), n.burnin=1000, n.sim=2000, n.thin=2)
geweke_plotTAR(fit2)

###### Example 3: Temperature, precipitation, and two river flows in Iceland
data(iceland.rf)
fit3 <- mtar(~ Jokulsa + Vatnsdalsa | Temperature | Precipitation,
             data=iceland.rf, subset={Date<="1974-11-06"}, row.names=Date,
             ars=ars(nregim=2,p=15,q=4,d=2), n.burnin=1000, n.sim=2000,
             n.thin=2, dist="Slash")
geweke_plotTAR(fit3)

###### Example 4: U.S. stock returns
data(US.returns)
fit4 <- mtar(~ CCR | dVIX, data=US.returns, subset={Date<="2025-11-28"},
             row.names=Date, ars=ars(nregim=2,p=3,d=3), n.burnin=1000,
             n.sim=2000, n.thin=2, dist="Student-t")
geweke_plotTAR(fit4)

}
}
\seealso{
\code{\link[coda]{geweke.plot}}
}
