% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesians.R
\name{HPDinterval.mtar}
\alias{HPDinterval.mtar}
\title{Highest Posterior Density intervals for objects of class \code{mtar}}
\usage{
\method{HPDinterval}{mtar}(obj, prob = 0.95, ...)
}
\arguments{
\item{obj}{an object of class \code{mtar} generated by a call to the function \code{mtar()}.}

\item{prob}{a numeric scalar in the interval \eqn{(0,1)} giving the target probability content of
the intervals. By default, \code{prob} is set to \code{0.95}.}

\item{...}{Optional additional arguments for methods. None are used at present.}
}
\description{
Highest Posterior Density intervals for objects of class \code{mtar}
}
\examples{
\donttest{
###### Example 1: Returns of the closing prices of three financial indexes
data(returns)
fit1 <- mtar(~ COLCAP + BOVESPA | SP500, data=returns, row.names=Date,
             subset={Date<="2015-12-07"}, dist="Student-t",
             ars=ars(nregim=3,p=c(1,1,2)), n.burnin=1000, n.sim=2000,
             n.thin=2, ssvs=TRUE)
coda::HPDinterval(fit1)

###### Example 2: Rainfall and two river flows in Colombia
data(riverflows)
fit2 <- mtar(~ Bedon + LaPlata | Rainfall, data=riverflows, row.names=Date,
             subset={Date<="2009-02-13"}, dist="Laplace",
             ars=ars(nregim=3,p=5), n.burnin=1000, n.sim=2000, n.thin=2)
coda::HPDinterval(fit2)

###### Example 3: Temperature, precipitation, and two river flows in Iceland
data(iceland.rf)
fit3 <- mtar(~ Jokulsa + Vatnsdalsa | Temperature | Precipitation,
             data=iceland.rf, subset={Date<="1974-11-06"}, row.names=Date,
             ars=ars(nregim=2,p=15,q=4,d=2), n.burnin=1000, n.sim=2000,
             n.thin=2, dist="Slash")
coda::HPDinterval(fit3)

###### Example 4: U.S. stock returns
data(US.returns)
fit4 <- mtar(~ CCR | dVIX, data=US.returns, subset={Date<="2025-11-28"},
             row.names=Date, ars=ars(nregim=2,p=3,d=3), n.burnin=1000,
             n.sim=2000, n.thin=2, dist="Student-t")
coda::HPDinterval(fit4)
}

}
\seealso{
\code{\link[coda]{HPDinterval}}
}
