% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesians.R
\name{out.of.sample}
\alias{out.of.sample}
\title{Out-of-sample predictive accuracy measures}
\usage{
out.of.sample(..., newdata, n.ahead)
}
\arguments{
\item{...}{one or several objects of the same class.}

\item{newdata}{a data frame containing the future values of the threshold series, if any in the
fitted model; the exogenous series, if any in the fitted model; and the output series.}

\item{n.ahead}{an integer value specifying the number of forecast steps}
}
\description{
Out-of-sample predictive accuracy measures
}
