% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions.R
\name{priors}
\alias{priors}
\title{Auxiliary function to set hyperparameter values}
\usage{
priors(prior, regim, k, dist, setar, ssvs)
}
\arguments{
\item{prior}{list of values for the hyperparameters. Hyperparameters not included in this list take
their default values.}

\item{regim}{positive integer value indicating the number of regimes.}

\item{k}{positive integer value indicating the dimension of the output series.}

\item{dist}{character string specifying the distribution used to describe the noise process behavior.}

\item{setar}{NULL if the model is SETAR, and a positive integer indicating the component of the output
series that acts as the threshold variable, otherwise.}

\item{ssvs}{a logical variable indicating if Stochastic Search Variable Selection (SSVS) is performed.}
}
\value{
a list composed of the hyperparameter values for the a priori distribution of all the model parameters.
}
\description{
Auxiliary function to set hyperparameter values
}
