% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simtar}
\alias{simtar}
\title{Simulation of multivariate time series according to a TAR model}
\usage{
simtar(
  n,
  k = 2,
  ars = ars(),
  Intercept = TRUE,
  trend = c("none", "linear", "quad"),
  nseason = NULL,
  parms,
  delay = 0,
  thresholds = NULL,
  t.series = NULL,
  ex.series = NULL,
  dist = c("Gaussian", "Student-t", "Hyperbolic", "Laplace", "Slash",
    "Contaminated normal", "Skew-Student-t", "Skew-normal"),
  skewness = NULL,
  extra = NULL,
  setar = NULL
)
}
\arguments{
\item{n}{a positive integer value indicating the length of the desired output series.}

\item{k}{a positive integer value indicating the dimension of the desired output series.}

\item{ars}{It is a list composed of four objects, namely: the number of regimes (\code{nregim}),
the autoregressive order (\code{p}), and the maximum lags for the exogenous (\code{q})
and the threshold (\code{d}) series within each regime. It can be validated using the
\code{ars()} routine.}

\item{Intercept}{an (optional) logical variable. If \code{TRUE}, then the model includes an intercept.}

\item{trend}{It is an (optional) character string that allows the user to specify the
degree of deterministic time trend to be included in each regime. The available
options are: linear trend ("linear"), quadratic trend ("quad"), and no time
trend ("none"). By default, \code{trend} is set to "none".}

\item{nseason}{It is an (optional) integer value that allows the user to specify the number
of seasonal periods. When the \code{nseason} is specified, \code{nseason}-1
seasonal dummies are added to the regressors within each regime.}

\item{parms}{a list with as many sublists as regimes in the user-specified TAR model. Each sublist is composed of two matrices. The first corresponds
to location parameters, while the second corresponds to scale parameters.}

\item{delay}{an (optional) non-negative integer value indicating the delay in the threshold series.}

\item{thresholds}{a vector with \eqn{l-1} real values sorted ascendingly.}

\item{t.series}{a matrix with the values of the threshold series.}

\item{ex.series}{a matrix with the values of the multivariate exogenous series.}

\item{dist}{an (optional) character string which allows the user to specify the multivariate
distribution to be used to describe the behavior of the noise process. The
available options are: Gaussian ("Gaussian"), Student-\eqn{t} ("Student-t"),
Slash ("Slash"), Symmetric Hyperbolic ("Hyperbolic"), Laplace ("Laplace"), and
contaminated normal ("Contaminated normal"). By default,\code{dist} is set to
"Gaussian".}

\item{skewness}{an (optional) vector with the values of the skewness parameters. By default,\code{delta} is set to \code{NULL}.}

\item{extra}{a value indicating the value of the extra parameter of the noise process distribution, if any.}

\item{setar}{an (optional) positive integer indicating the component of the output series which should
be the threshold variable. By default,\code{setar} is set to \code{NULL}.}
}
\value{
a \code{data.frame} containing the output series, threshold series (if any), and multivariate exogenous series (if any).
}
\description{
This function simulates multivariate time series according to a user-specified TAR model.
}
\examples{
\donttest{
###### Simulation of a trivariate TAR model with two regimes
n <- 2000
k <- 3
myars <- ars(nregim=2,p=c(1,2))
Z <- as.matrix(arima.sim(n=n+max(myars$p),list(ar=c(0.5))))
probs <- sort((0.6 + runif(myars$nregim-1)*0.8)*c(1:(myars$nregim-1))/myars$nregim)
parms <- list()
for(j in 1:myars$nregim){
    np <- 1 + myars$p[j]*k
    parms[[j]] <- list()
    parms[[j]]$location <- c(ifelse(runif(np*k)<=0.5,1,-1)*rbeta(np*k,shape1=4,shape2=16))
    parms[[j]]$location <- matrix(parms[[j]]$location,np,k)
    parms[[j]]$scale <- rgamma(k,shape=1,scale=1)*diag(k)
}
thresholds <- quantile(Z,probs=probs)
out1 <- simtar(n=n,k=k,ars=myars,parms=parms,thresholds=thresholds,
               t.series=Z,dist="Student-t",extra=6)
str(out1)

fit1 <- mtar(~ Y1 + Y2 + Y3 | t.series, data=out1, ars=myars, dist="Student-t",
             n.sim=3000, n.burn=2000, n.thin=2)
summary(fit1)

###### Simulation of a trivariate VAR model
n <- 2000
k <- 3
myars <- ars(nregim=1,p=2)
parms <- list()
for(j in 1:myars$nregim){
    np <- 1 + myars$p[j]*k
    parms[[j]] <- list()
    parms[[j]]$location <- c(ifelse(runif(np*k)<=0.5,1,-1)*rbeta(np*k,shape1=4,shape2=16))
    parms[[j]]$location <- matrix(parms[[j]]$location,np,k)
    parms[[j]]$scale <- rgamma(k,shape=1,scale=1)*diag(k)
}
out2 <- simtar(n=n,k=k,ars=myars,parms=parms,dist="Slash",extra=2)
str(out2)

fit2 <- mtar(~ Y1 + Y2 + Y3, data=out2, ars=myars, dist="Slash", n.sim=3000,
             n.burn=2000, n.thin=2)
summary(fit2)

n <- 5000
k <- 3
myars <- ars(nregim=2,p=c(1,2))
parms <- list()
for(j in 1:myars$nregim){
    np <- 1 + myars$p[j]*k
    parms[[j]] <- list()
    parms[[j]]$location <- c(ifelse(runif(np*k)<=0.5,1,-1)*rbeta(np*k,shape1=4,shape2=16))
    parms[[j]]$location <- matrix(parms[[j]]$location,np,k)
    parms[[j]]$scale <- rgamma(k,shape=1,scale=1)*diag(k)
}
out3 <- simtar(n=n, k=k, ars=myars, parms=parms, delay=2,
               thresholds=-1, dist="Laplace", setar=2)
str(out3)

fit3 <- mtar(~ Y1 + Y2 + Y3, data=out3, ars=myars, dist="Laplace",
             n.sim=3000, n.burn=2000, n.thin=2, setar=2)
summary(fit3)

}

}
\references{
Vanegas, L.H. and Calderón, S.A. and Rondón, L.M. (2025) Bayesian estimation of a multivariate tar model when the
            noise process distribution belongs to the class of gaussian variance mixtures. International Journal
            of Forecasting.
}
