% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{vcov.mtar}
\alias{vcov.mtar}
\title{vcov method for objects of class \code{mtar}}
\usage{
\method{vcov}{mtar}(object, ..., FUN = mean)
}
\arguments{
\item{object}{an object of class \code{mtar} generated by a call to the function \code{mtar()}.}

\item{...}{optional arguments to FUN.}

\item{FUN}{the function to be applied to the chains for the scale parameters resulting
from the MCMC-type algorithm. By default, \code{FUN} is set to \code{mean()}.}
}
\value{
a list composed of the values obtained from \code{FUN} when the latter is computed using
        chains resulting from the MCMC-type algorithm for the scale parameters.
}
\description{
vcov method for objects of class \code{mtar}
}
