% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesians.R
\name{WAIC}
\alias{WAIC}
\title{Watanabe-Akaike or Widely Available Information Criterion (WAIC)}
\usage{
WAIC(...)
}
\arguments{
\item{...}{one or several objects of the same class.}
}
\value{
A matrix with the values of the WAIC for each object in the input.
}
\description{
This function computes the Watanabe-Akaike or Widely Available Information Criterion (WAIC), which is an adjusted within-sample predictive accuracy measure.
}
\references{
Watanabe S. (2010). Asymptotic Equivalence of Bayes Cross Validation and Widely Applicable Information Criterion in
            Singular Learning Theory. The Journal of Machine Learning Research, 11, 3571–3594.
}
