% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesians.R
\name{as.mcmc.mtar}
\alias{as.mcmc.mtar}
\title{Coercion of \code{mtar} objects to \code{mcmc} objects}
\usage{
\method{as.mcmc}{mtar}(x, ...)
}
\arguments{
\item{x}{an object of class \code{mtar} obtained from a call to \code{mtar()}.}

\item{...}{additional arguments passed to specific coercion methods.}
}
\value{
A list of \code{mcmc} objects containing the posterior simulation draws
generated by the \code{mtar()} routine.
}
\description{
This method converts an object of class \code{mtar} into a list of
\code{mcmc} objects, each corresponding to a Markov chain produced during
Bayesian estimation.
}
\examples{
\donttest{
###### Example 1: Returns of the closing prices of three financial indexes
data(returns)
fit1 <- mtar(~ COLCAP + BOVESPA | SP500, data=returns, row.names=Date,
             subset={Date<="2015-12-07"}, dist="Student-t",
             ars=ars(nregim=3,p=c(1,1,2)), n.burnin=100, n.sim=200,
             n.thin=2, ssvs=TRUE)
fit1.mcmc <- coda::as.mcmc(fit1)
summary(fit1.mcmc)
#plot(fit1.mcmc)

###### Example 2: Rainfall and two river flows in Colombia
data(riverflows)
fit2 <- mtar(~ Bedon + LaPlata | Rainfall, data=riverflows, row.names=Date,
             subset={Date<="2009-02-13"}, dist="Laplace",
             ars=ars(nregim=3,p=5), n.burnin=100, n.sim=200, n.thin=2)
fit2.mcmc <- coda::as.mcmc(fit2)
summary(fit2.mcmc)
#plot(fit2.mcmc)

###### Example 3: Temperature, precipitation, and two river flows in Iceland
data(iceland.rf)
fit3 <- mtar(~ Jokulsa + Vatnsdalsa | Temperature | Precipitation,
             data=iceland.rf, subset={Date<="1974-11-06"}, row.names=Date,
             ars=ars(nregim=2,p=15,q=4,d=2), n.burnin=100, n.sim=200,
             n.thin=2, dist="Slash")
fit3.mcmc <- coda::as.mcmc(fit3)
summary(fit3.mcmc)
#plot(fit3.mcmc)

###### Example 4: U.S. stock returns
data(US.returns)
fit4 <- mtar(~ CCR | dVIX, data=US.returns, subset={Date<="2025-11-28"},
             row.names=Date, ars=ars(nregim=2,p=3,d=3), n.burnin=100,
             n.sim=200, n.thin=2, dist="Student-t")
fit4.mcmc <- coda::as.mcmc(fit4)
summary(fit4.mcmc)
#plot(fit4.mcmc)
}


}
\seealso{
\code{\link[coda]{as.mcmc}}
}
