% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesians.R
\name{as.mcmc.mtar}
\alias{as.mcmc.mtar}
\title{Method to coerce objects of class \code{mtar} to \code{mcmc} objects.}
\usage{
\method{as.mcmc}{mtar}(x, ...)
}
\arguments{
\item{x}{an object of class \code{mtar} generated by a call to the function \code{mtar()}.}

\item{...}{Further arguments to be passed to specific methods.}
}
\value{
A list composed of objects of class \code{mcmc} that correspond to
        chains resulting from a call to the routine \code{mtar()}.
}
\description{
Method to coerce objects of class \code{mtar} to \code{mcmc} objects.
}
\examples{
\donttest{
###### Example 1: Returns of the closing prices of three financial indexes
data(returns)
fit1 <- mtar(~ COLCAP + BOVESPA | SP500, data=returns, row.names=Date,
             dist="Gaussian", ars=ars(nregim=3,p=c(1,1,2)), n.burnin=100,
             n.sim=3000, n.thin=2)
chains1 <- coda::as.mcmc(fit1)
summary(chains1)

###### Example 2: Rainfall and two river flows in Colombia
data(riverflows)
fit2 <- mtar(~ Bedon + LaPlata | Rainfall, data=riverflows, row.names=Date,
             dist="Gaussian", ars=ars(nregim=3,p=5), n.burnin=2000,
             n.sim=3000, n.thin=2)
chains2 <- coda::as.mcmc(fit2)
summary(chains2)
}
}
