% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{coef.mtar}
\alias{coef.mtar}
\title{coef method for objects of class \code{mtar}}
\usage{
\method{coef}{mtar}(object, ..., FUN = mean)
}
\arguments{
\item{object}{an object of class \code{mtar} obtained from a call to the \code{mtar()} function.}

\item{...}{additional arguments passed to \code{FUN}.}

\item{FUN}{a function to be applied to the MCMC chains associated with each model parameter.
By default, \code{FUN} is set to \code{mean}.}
}
\value{
A list containing the summary statistics obtained by applying \code{FUN} to the
MCMC chains of each model parameter.
}
\description{
coef method for objects of class \code{mtar}
}
