% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesians.R
\name{HPDinterval.mtar}
\alias{HPDinterval.mtar}
\title{Highest Posterior Density intervals for objects of class \code{mtar}}
\usage{
\method{HPDinterval}{mtar}(obj, prob = 0.95, ...)
}
\arguments{
\item{obj}{an object of class \code{mtar} generated by a call to the function \code{mtar()}.}

\item{prob}{a numeric scalar in the interval \eqn{(0,1)} giving the target probability content of
the intervals. By default, \code{prob} is set to \code{0.95}.}

\item{...}{Optional additional arguments for methods. None are used at present.}
}
\description{
Highest Posterior Density intervals for objects of class \code{mtar}
}
\seealso{
\code{\link[coda]{HPDinterval}}
}
