% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{mtar}
\alias{mtar}
\title{Bayesian estimation of a multivariate Threshold Autoregressive (TAR) model.}
\usage{
mtar(
  formula,
  data,
  subset,
  Intercept = TRUE,
  trend = c("none", "linear", "quad"),
  nseason = NULL,
  ars = ars(),
  row.names,
  dist = c("Gaussian", "Student-t", "Hyperbolic", "Laplace", "Slash",
    "Contaminated normal", "Skew-Student-t", "Skew-normal"),
  prior = list(),
  n.sim = 500,
  n.burnin = 100,
  n.thin = 1,
  ssvs = FALSE,
  setar = NULL,
  log = FALSE,
  ...
)
}
\arguments{
\item{formula}{a three-part expression of type \code{Formula} describing the TAR model
to be fitted to the data. In the first part, the variables in the
multivariate output series are listed; in the second part, the threshold
series is specified, and in the third part, the variables in the
multivariate exogenous series are specified.}

\item{data}{a data frame containing the variables in the model.
If not found in data, the variables are taken from \code{environment(formula)},
typically the environment from which \code{mtar} is called.}

\item{subset}{an (optional) vector specifying a subset of observations to be used in the
fitting process.}

\item{Intercept}{an (optional) logical variable. If \code{TRUE}, then the model
includes an intercept.}

\item{trend}{an (optional) character string that allows the user to specify the
degree of deterministic time trend to be included in each regime. The available
options are: linear trend ("linear"), quadratic trend ("quad"), and no time
trend ("none"). By default, \code{trend} is set to "none".}

\item{nseason}{an (optional) integer value that allows the user to specify the number
of seasonal periods. When the \code{nseason} is specified, \code{nseason}-1
seasonal dummies are added to the regressors within each regime.}

\item{ars}{a list composed of four objects, namely: the number of regimes (\code{nregim}),
the autoregressive order (\code{p}), and the maximum lags for the exogenous (\code{q})
and the threshold (\code{d}) series within each regime. It can be validated using the
routine \code{ars()}. Only \code{nregim} and \code{p} are required.}

\item{row.names}{an (optional) vector that allows the user to name the time point to
which each row in the data set corresponds.}

\item{dist}{an (optional) character string that allows the user to specify the
multivariate distribution to be used to describe the noise process
behavior. The available options are: Gaussian ("Gaussian"), Student-\eqn{t}
("Student-t"), Slash ("Slash"), Symmetric Hyperbolic ("Hyperbolic"),
Laplace ("Laplace"), Contaminated normal ("Contaminated normal"),
Skew-normal ("Skew-normal") and Skew-Student-\eqn{t} ("Skew-Student-t").
By default, \code{dist} is set to "Gaussian".}

\item{prior}{an (optional) list that allows the user to specify the values of the
hyperparameters, that is, allows to specify the values of the parameters
of the prior distributions. It can be validated using the \code{priors()} routine.}

\item{n.sim}{an (optional) positive integer specifying the required number of iterations
for the simulation after the burn-in period. By default, \code{n.sim} is set
to 500.}

\item{n.burnin}{an (optional) positive integer specifying the required number of burn-in
iterations for the simulation. By default, \code{n.burnin} is set to 100.}

\item{n.thin}{an (optional) positive integer specifying the required thinning interval
for the simulation. By default, \code{n.thin} is set to 1.}

\item{ssvs}{an (optional) logical variable. If \code{ssvs=TRUE} then the Stochastic Search
Variable Selection (SSVS) procedure is applied to identify relevant lags of the
output, exogenous and threhold series. By default, \code{ssvs} is set to \code{FALSE}.}

\item{setar}{an (optional) positive integer indicating the component of the output series which is
the threshold variable. By default,\code{setar} is set to \code{NULL}, which
indicates that the fitted model is not SETAR.}

\item{log}{an (optional) logical variable. If \code{TRUE}, then the behaviour of the output
series is described using the exponentiated version of \code{dist}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
an object of class \emph{mtar} in which the main results of the model fitted to the data are stored, i.e., a
list with components including
\tabular{ll}{
\code{chains}   \tab list with several arrays, which store the values of each model parameter in each iteration of the simulation,\cr
\tab \cr
\code{n.sim}    \tab number of iterations of the simulation after the burn-in period,\cr
\tab \cr
\code{n.burnin} \tab number of burn-in iterations in the simulation,\cr
\tab \cr
\code{n.thin}   \tab thinning interval in the simulation,\cr
\tab \cr
\code{ars}      \tab list composed of four objects, namely: \code{nregim}, \code{p}, \code{q} and \code{d},
                     each of which corresponds to a vector of non-negative integers with as
                     many elements as there are regimes in the fitted TAR model,\cr
\tab \cr
\code{dist}     \tab name of the multivariate distribution used to describe the behavior of
                     the noise process,\cr
\tab \cr
\code{threshold.series}  \tab vector with the values of the threshold series,\cr
\tab \cr
\code{output.series}   \tab matrix with the values of the output series,\cr
\tab \cr
\code{exogenous.series} \tab matrix with the values of the exogenous series,\cr
\tab \cr
\code{Intercept}    \tab If \code{TRUE}, then the model included an intercept term in each regime,\cr
\tab \cr
\code{trend}    \tab the degree of the deterministic time trend, if any,\cr
\tab \cr
\code{nseason}    \tab the number of seasonal periods, if any,\cr
\tab \cr
\code{formula}      \tab the formula,\cr
\tab \cr
\code{call}         \tab the original function call.\cr
}
}
\description{
This function uses Gibbs sampling to generate a sample from the posterior
             distribution of the parameters of a multivariate TAR model when the noise
             process follows Gaussian, Student-\eqn{t}, Slash, Symmetric Hyperbolic,
             Contaminated normal, Laplace, Skew-normal, or skew-\eqn{t} distribution.
}
\examples{
\donttest{
###### Example 1: Returns of the closing prices of three financial indexes
data(returns)
fit1 <- mtar(~ COLCAP + BOVESPA | SP500, data=returns, row.names=Date,
             dist="Gaussian", ars=ars(nregim=3,p=c(1,1,2)), n.burnin=100,
             n.sim=3000, n.thin=2)
summary(fit1)

###### Example 2: Rainfall and two river flows in Colombia
data(riverflows)
fit2 <- mtar(~ Bedon + LaPlata | Rainfall, data=riverflows, row.names=Date,
             dist="Gaussian", ars=ars(nregim=3,p=5), n.burnin=2000,
             n.sim=3000, n.thin=2)
summary(fit2)
}


}
\references{
Nieto, F.H. (2005) Modeling Bivariate Threshold Autoregressive Processes in the Presence of Missing Data.
            Communications in Statistics - Theory and Methods, 34, 905-930.

Romero, L.V. and Calderon, S.A. (2021) Bayesian estimation of a multivariate TAR model when the noise
            process follows a Student-t distribution. Communications in Statistics - Theory and Methods, 50, 2508-2530.

Calderon, S.A. and Nieto, F.H. (2017) Bayesian analysis of multivariate threshold autoregressive models
            with missing data. Communications in Statistics - Theory and Methods, 46, 296-318.

Vanegas, L.H. and Calderón, S.A. and Rondón, L.M. (2025) Bayesian estimation of a multivariate tar model when the
            noise process distribution belongs to the class of gaussian variance mixtures. International Journal
            of Forecasting.
}
\seealso{
\link{DIC}, \link{WAIC}
}
