% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperfunctions.R
\name{mtar_grid}
\alias{mtar_grid}
\title{Bayesian estimation of a multivariate TAR model}
\usage{
mtar_grid(
  formula,
  data,
  subset,
  Intercept = TRUE,
  trend = c("none", "linear", "quad"),
  nseason = NULL,
  nregim.min = 1,
  nregim.max = NULL,
  p.min = 1,
  p.max = NULL,
  q = FALSE,
  d = FALSE,
  row.names,
  dist = "Gaussian",
  prior = list(),
  n.sim = 500,
  n.burnin = 100,
  n.thin = 1,
  ssvs = FALSE,
  setar = NULL,
  log = FALSE
)
}
\arguments{
\item{formula}{a three-part expression of type \code{Formula} describing the TAR model
to be fitted to the data. In the first part, the variables in the
multivariate output series are listed; in the second part, the threshold
series is specified, and in the third part, the variables in the
multivariate exogenous series are specified.}

\item{data}{a data frame containing the variables in the model.
If not found in data, the variables are taken from \code{environment(formula)},
typically the environment from which \code{mtar_grid()} is called.}

\item{subset}{an (optional) vector specifying a subset of observations to be used in the
fitting process.}

\item{Intercept}{an (optional) logical variable. If \code{TRUE} then the model
includes an intercept.}

\item{trend}{an (optional) character string that allows the user to specify the
degree of deterministic time trend to be included in each regime. The available
options are: linear trend ("linear"), quadratic trend ("quad"), and no time
trend ("none"). By default, \code{trend} is set to "none".}

\item{nseason}{an (optional) integer value that allows the user to specify the number
of seasonal periods. When the \code{nseason} is specified, \code{nseason}-1
seasonal dummies are added to the regressors within each regime.}

\item{nregim.min}{an (optional) integer that represents the minimum value for the number
of regimes. By default, \code{nregim.min} is set to 1.}

\item{nregim.max}{an integer that represents the maximum value for the number
of regimes.}

\item{p.min}{an (optional) integer that represents the minimum value for the autoregressive
order within each regime. By default, \code{p.min} is set to 1.}

\item{p.max}{an integer that represents the maximum value for the autoregressive
order within each regime.}

\item{q}{an (optional) logical variable. If \code{TRUE} then the maximum lag for the exogenous
series within each regime is set to be equal to the autoregressive order. By default,
\code{q} is set to FALSE.}

\item{d}{an (optional) logical variable. If \code{TRUE} then the maximum lag for the threshold
series within each regime is set to be equal to the autoregressive order. By default,
\code{d} is set to FALSE.}

\item{row.names}{an (optional) vector that allows the user to name the time point to
which each row in the data set corresponds.}

\item{dist}{Specifies the multivariate distributions to be used to describe the noise process
behavior as a character string vector. The available options are: Gaussian
("Gaussian"), Student-\eqn{t} ("Student-t"), Slash ("Slash"), Symmetric Hyperbolic ("Hyperbolic"),
Laplace ("Laplace"), Contaminated normal ("Contaminated normal"),
Skew-normal ("Skew-normal") and Skew-Student-\eqn{t} ("Skew-Student-t").
By default, \code{dist} is set to "Gaussian".}

\item{prior}{an (optional) list that allows the user to specify the values of the
hyperparameters, that is, allows to specify the values of the parameters
of the prior distributions. It can be validated using the \code{priors()} routine.}

\item{n.sim}{an (optional) positive integer specifying the required number of iterations
for the simulation after the burn-in period. By default, \code{n.sim} is set
to 500.}

\item{n.burnin}{an (optional) positive integer specifying the required number of burn-in
iterations for the simulation. By default, \code{n.burnin} is set to 100.}

\item{n.thin}{an (optional) positive integer specifying the required thinning interval
for the simulation. By default, \code{n.thin} is set to 1.}

\item{ssvs}{an (optional) logical variable. If \code{ssvs=TRUE} then the Stochastic Search
Variable Selection (SSVS) procedure is applied to identify relevant lags of the
output, exogenous and threhold series. By default, \code{ssvs} is set to \code{FALSE}.}

\item{setar}{an (optional) positive integer indicating the component of the output series which is
the threshold variable. By default,\code{setar} is set to \code{NULL}, which
indicates that the fitted model is not SETAR.}

\item{log}{an (optional) logical variable. If \code{TRUE} then the behaviour of the output
series is described using the exponentiated version of \code{dist}.}
}
\value{
a list-type object comprising objects of class \code{mtar}, which correspond to the
results of a call to the routine \code{mtar()} for each combination of the noise process distribution
 (\code{dist}), the number of regimes (\code{nregim.min} to \code{nregim.max}) and the autoregressive
 order within each regime (\code{p.min} to \code{p.max}).
}
\description{
This is a wrapper function that applies the routine \code{mtar()} for each combination
of the noise process distribution (\code{dist}), the number of regimes (\code{nregim.min} to \code{nregim.max})
and the autoregressive order within each regime (\code{p.min} to \code{p.max}). In all calls to \code{mtar()},
the time points employed in the fitting process are the same, which is accomplished by managing the
argument \code{subset} of \code{mtar()}.
}
\seealso{
mtar
}
