% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesians.R
\name{out.of.sample}
\alias{out.of.sample}
\title{Out-of-sample predictive accuracy measures}
\usage{
out.of.sample(..., newdata, n.ahead)
}
\arguments{
\item{...}{one or more fitted model objects of the same class.}

\item{newdata}{a data frame containing the future values of the output series required to evaluate
predictive performance.}

\item{n.ahead}{a positive integer specifying the number of forecast steps ahead to use in the
predictive performance evaluation.}
}
\description{
Computes out-of-sample predictive accuracy measures for one or more fitted models
of the same class, based on their predictive distributions.
}
