% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.R
\name{out_of_sample.listmtar}
\alias{out_of_sample.listmtar}
\title{Computing Out-of-Sample predictive accuracy measures}
\usage{
\method{out_of_sample}{listmtar}(
  x,
  newdata,
  n.ahead = NULL,
  credible = 0.95,
  by.component = FALSE,
  FUN = mean,
  ...
)
}
\arguments{
\item{x}{An object of class \code{listmtar} returned by the routine \code{mtar_grid()}.}

\item{newdata}{A \code{data.frame} containing future values of the threshold
series (if included in the fitted model), the exogenous series (if included in the fitted
model), and the realized values of the output series.}

\item{n.ahead}{A positive integer specifying the number of steps ahead to forecast.}

\item{credible}{An optional numeric value in \eqn{(0,1)} specifying the level of the
required prediction intervals. By default, \code{credible} is set to \code{0.95}.}

\item{by.component}{An optional logical argument. If \code{TRUE}, the predictive
accuracy measures are computed separately for each component of the multivariate
output series. By default, \code{by.component} is set to \code{TRUE}.}

\item{FUN}{An optional function used to summarize the \code{n.ahead} values computed
for each predictive accuracy measure. By default, \code{FUN} is set to \code{mean}.}

\item{...}{optional arguments to FUN.}
}
\description{
Computes Out-of-Sample predictive accuracy measures for an object of class \code{listmtar}.
}
\examples{
\donttest{
###### Example 1: Returns of the closing prices of three financial indexes
data(returns)
fit1 <- mtar_grid(~ COLCAP + BOVESPA | SP500, data=returns, row.names=Date,
                  subset={Date<="2015-12-07"}, dist=c("Gaussian","Student-t",
                  "Slash","Laplace"), nregim.min=2, nregim.max=3, p.min=2,
                  p.max=2, n.burnin=100, n.sim=200, n.thin=2,
                  plan_strategy="multisession")
oos1 <- out_of_sample(fit1, newdata=subset(returns, Date>"2015-12-07"),
                      n.ahead=75, by.component=TRUE, FUN=median)
oos1

###### Example 2: Rainfall and two river flows in Colombia
data(riverflows)
fit2 <- mtar_grid(~ Bedon + LaPlata | Rainfall, data=riverflows,
                  row.names=Date, subset={Date<="2009-02-13"},dist="Laplace",
                  nregim.min=2, nregim.max=3, p.min=1, p.max=3,n.burnin=100,
                  n.sim=200, n.thin=2, plan_strategy="multisession")
oos2 <- out_of_sample(fit2, newdata=subset(riverflows, Date>"2009-02-13"),
                      n.ahead=60, by.component=TRUE, FUN=median)
oos2

###### Example 3: Temperature, precipitation, and two river flows in Iceland
data(iceland.rf)
fit3 <- mtar_grid(~ Jokulsa + Vatnsdalsa | Temperature | Precipitation,
                  data=iceland.rf,subset={Date<="1974-11-06"},row.names=Date,
                  dist=c("Slash","Student-t"), nregim.min=1, nregim.max=2,
                  p.min=15, p.max=15, q.min=4, q.max=4, d.min=2, d.max=2,
                  n.burnin=100, n.sim=200, n.thin=2,
                  plan_strategy="multisession")
oos3 <- out_of_sample(fit3, newdata=subset(iceland.rf, Date>"1974-11-06"),
                      n.ahead=55, by.component=TRUE, FUN=median)
oos3

###### Example 4: U.S. stock returns
data(US.returns)
fit4 <- mtar_grid(~ CCR | dVIX, data=US.returns, subset={Date<="2025-11-28"},
                  row.names=Date, dist=c("Laplace","Student-t","Slash"),
                  nregim.min=2, nregim.max=2, p.min=3, p.max=3, d.min=3,
                  d.max=3, n.burnin=100, n.sim=200, n.thin=2,
                  plan_strategy="multisession")
oos4 <- out_of_sample(fit4, newdata=subset(US.returns, Date>"2025-11-28"),
                      n.ahead=100, by.component=TRUE, FUN=median)
oos4
}


}
