% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.R
\name{out_of_sample.mtar}
\alias{out_of_sample.mtar}
\title{Computing Out-of-Sample predictive accuracy measures}
\usage{
\method{out_of_sample}{mtar}(
  ...,
  newdata,
  n.ahead = NULL,
  credible = 0.95,
  by.component = TRUE,
  FUN = mean
)
}
\arguments{
\item{...}{one or more objects of class \code{mtar}.}

\item{newdata}{A \code{data.frame} containing future values of the threshold
series (if included in the fitted model), the exogenous series (if included in the fitted
model), and the realized values of the output series.}

\item{n.ahead}{A positive integer specifying the number of steps ahead to forecast.}

\item{credible}{An optional numeric value in \eqn{(0,1)} specifying the level of the
required prediction intervals. By default, \code{credible} is set to \code{0.95}.}

\item{by.component}{An optional logical argument. If \code{TRUE}, the predictive
accuracy measures are computed separately for each component of the multivariate
output series. By default, \code{by.component} is set to \code{TRUE}.}

\item{FUN}{An optional function used to summarize the \code{n.ahead} values computed
for each predictive accuracy measure. By default, \code{FUN} is set to \code{mean}.}
}
\description{
Computes Out-of-Sample predictive accuracy measures for two or more objects of class \code{mtar}.
}
