% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.R
\name{predict.mtar}
\alias{predict.mtar}
\title{Forecasting for a multivariate TAR model.}
\usage{
\method{predict}{mtar}(
  object,
  ...,
  newdata,
  n.ahead = 1,
  row.names,
  credible = 0.95,
  out.of.sample = FALSE
)
}
\arguments{
\item{object}{an object of the class \emph{mtar}.}

\item{...}{additional arguments affecting the predictions produced}

\item{newdata}{an (optional) data frame containing the future values of the
threshold series, if any in the fitted model; the exogenous
series, if any in the fitted model; and the output series
provided that \code{out.of.sample=TRUE}.}

\item{n.ahead}{an integer value specifying the number of forecast steps.}

\item{row.names}{an (optional) vector that allows the user to name the time point to
which each row in \code{newdata} corresponds}

\item{credible}{an (optional) value for the level of the credible intervals. By default, \code{credible} is set to 0.95.}

\item{out.of.sample}{an (optional) logical variable. If \code{TRUE}, then the log-score is computed, which is a measurement to assess density forecasts. Therefore, the data.frame specified in the argument \code{data} must to include the true values of the output series.}
}
\value{
a list with the following component

\tabular{ll}{
\code{ypred}   \tab a matrix with the results of the forecasting,\cr
\tab \cr
\code{summary} \tab a matrix with the mean and credible intervals of the forecasting,\cr
}
}
\description{
This function computes forecasting from a fitted multivariate TAR model.
}
\examples{
\donttest{
###### Example 1: Returns of the closing prices of three financial indexes
data(returns)
fit1 <- mtar(~ COLCAP + BOVESPA | SP500, data=returns, subset={Date<="2016-03-14"},
             row.names=Date, dist="Gaussian", ars=ars(nregim=3,p=c(1,1,2)), n.burnin=100,
             n.sim=3000, n.thin=2)
out1 <- predict(fit1, newdata=subset(returns,Date>"2016-03-14"), n.ahead=10)
out1$summary

###### Example 2: Rainfall and two river flows in Colombia
data(riverflows)
fit2 <- mtar(~ Bedon + LaPlata | Rainfall, data=riverflows, subset={Date<="2009-04-04"},
             row.names=Date, dist="Gaussian", ars=ars(nregim=3,p=5), n.burnin=2000,
             n.sim=3000, n.thin=2)
out2 <- predict(fit2, newdata=subset(riverflows,Date>"2009-04-04"), n.ahead=10)
out2$summary
}
}
\references{
Nieto, F.H. (2005) Modeling Bivariate Threshold Autoregressive Processes in the Presence of Missing Data.
            Communications in Statistics - Theory and Methods, 34, 905-930.

Romero, L.V. and Calderon, S.A. (2021) Bayesian estimation of a multivariate TAR model when the noise
            process follows a Student-t distribution. Communications in Statistics - Theory and Methods, 50, 2508-2530.

Calderon, S.A. and Nieto, F.H. (2017) Bayesian analysis of multivariate threshold autoregressive models
            with missing data. Communications in Statistics - Theory and Methods, 46, 296-318.

Karlsson, S. (2013) Chapter 15-Forecasting with Bayesian Vector Autoregression. In Elliott, G. and
            Timmermann, A. Handbook of Economic Forecasting, Volume 2, 791–89, Elsevier.

Vanegas, L.H. and Calderón, S.A. and Rondón, L.M. (2025) Bayesian estimation of a multivariate tar model when the
            noise process distribution belongs to the class of gaussian variance mixtures. International Journal
            of Forecasting.
}
