% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{vcov.mtar}
\alias{vcov.mtar}
\title{vcov method for objects of class \code{mtar}}
\usage{
\method{vcov}{mtar}(object, ..., FUN = mean)
}
\arguments{
\item{object}{an object of class \code{mtar}, typically obtained from a call to
\code{mtar()}.}

\item{...}{additional arguments passed to \code{FUN}.}

\item{FUN}{a function to be applied to the MCMC chains of the scale parameters in order
to obtain point estimates. By default, \code{FUN} is set to \code{mean()}.}
}
\value{
A list containing the variance–covariance estimates obtained by applying
        \code{FUN} to the MCMC chains associated with the scale parameters.
}
\description{
Computes estimates of the variance–covariance matrices for the scale
parameters of a fitted multivariate TAR model.
}
