% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesians.R
\name{WAIC.mtar}
\alias{WAIC.mtar}
\title{Watanabe-Akaike or Widely Available Information Criterion (WAIC)}
\usage{
\method{WAIC}{mtar}(...)
}
\arguments{
\item{...}{one or several objects of the class \emph{mtar}.}
}
\value{
A matrix with the values of the WAIC for each \emph{mtar} object in the input.
}
\description{
This function computes the Watanabe-Akaike or Widely Available Information Criterion (WAIC), which is a within-sample predictive
             accuracy measure, for objects of class \code{mtar}.
}
\examples{
\donttest{
###### Example 1: Returns of the closing prices of three financial indexes
data(returns)
fit1a <- mtar(~ COLCAP + BOVESPA | SP500, data=returns, row.names=Date,
              dist="Gaussian", ars=ars(nregim=3,p=c(1,1,2)), n.burnin=100,
              n.sim=3000, n.thin=2)
fit1b <- update(fit1a,dist="Slash")
fit1c <- update(fit1a,dist="Student-t")
WAIC(fit1a,fit1b,fit1c)

###### Example 2: Rainfall and two river flows in Colombia
data(riverflows)
fit2a <- mtar(~ Bedon + LaPlata | Rainfall, data=riverflows, row.names=Date,
              dist="Gaussian", ars=ars(nregim=3,p=5), n.burnin=100,
              n.sim=3000, n.thin=2)
fit2b <- update(fit2a,dist="Slash")
fit2c <- update(fit2a,dist="Student-t")
WAIC(fit2a,fit2b,fit2c)
}

}
\seealso{
\link{DIC}
}
