% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesians.R
\name{WAIC}
\alias{WAIC}
\title{Watanabe-Akaike or Widely Available Information Criterion (WAIC)}
\usage{
WAIC(...)
}
\arguments{
\item{...}{one or more fitted model objects of the same class.}
}
\value{
A numeric matrix containing the WAIC values corresponding to each fitted object supplied in \code{...}.
}
\description{
Computes Watanabe-Akaike or Widely Available Information Criterion (WAIC), an adjusted
within-sample measure of predictive accuracy, for models estimated using Bayesian methods.
}
\references{
Watanabe S. (2010). Asymptotic Equivalence of Bayes Cross Validation and Widely Applicable Information Criterion in
            Singular Learning Theory. The Journal of Machine Learning Research, 11, 3571–3594.
}
